Holds that are ready now say "DUE." Account for this.
Holds that were ready to pick up used to say "Ready. Must be picked up by mm-dd-yy."
The library has recently switched to the more confusing "DUE mm-dd-yy" as if that's
the date the item is due back rather than the date by which you must pick it up.
Allow for either type of notification and return the same notice for each.