Changed logic for following links.
After logging in, the robot is taken to a different page depending on the
status of the account. If there are items checked out, the page has those
listed. If there are no items checked out but there are items on hold, the
page has the holds listed. If there are neither items checked out or on hold,
there's nothing to list.